now supports Alt+X to leave the app
